moved models/index to types/editor/index
