+++ apps/studio/electron/main/code/diff/insert.ts
@@ -3,7 +3,7 @@ import * as t from '@babel/types';
 import { parseJsxCodeBlock } from '../helpers';
 import { addKeyToElement, addUuidToElement, jsxFilter } from './helpers';
 import { assertNever } from '/common/helpers';
-import { InsertPos } from '/common/models';
+import { InsertPos } from '@onlook/types/editor';
 import type { CodeInsert } from '/common/models/actions/code';
 
 export function insertElementToNode(path: NodePath<t.JSXElement>, element: CodeInsert): void {
