+++ apps/studio/electron/main/code/diff/remove.ts
@@ -2,7 +2,7 @@ import type { NodePath } from '@babel/traverse';
 import type * as t from '@babel/types';
 import { jsxFilter } from './helpers';
 import { assertNever } from '/common/helpers';
-import { InsertPos } from '/common/models';
+import { InsertPos } from '@onlook/types/editor';
 import type { CodeAction } from '/common/models/actions/code';
 
 export function removeElementFromNode(path: NodePath<t.JSXElement>, element: CodeAction): void {
