+++ apps/studio/electron/preload/webview/elements/dom/insert.ts
@@ -2,7 +2,7 @@ import { cssManager } from '../../style';
 import { getDeepElement, getDomElement } from '../helpers';
 import { EditorAttributes, INLINE_ONLY_CONTAINERS } from '@onlook/types/constants';
 import { assertNever, getUniqueSelector } from '/common/helpers';
-import { InsertPos } from '/common/models';
+import { InsertPos } from '@onlook/types/editor';
 import type { ActionElement, ActionElementLocation } from '/common/models/actions';
 import type { DomElement } from '/common/models/element';
