+++ apps/studio/src/lib/editor/engine/move/index.ts
@@ -1,7 +1,7 @@
 import type React from 'react';
 import type { EditorEngine } from '..';
 import { escapeSelector } from '/common/helpers';
-import { InsertPos } from '/common/models';
+import { InsertPos } from '@onlook/types/editor';
 import type { MoveElementAction } from '/common/models/actions';
 import type { DomElement, ElementPosition } from '/common/models/element';
