+++ apps/studio/src/routes/editor/LayersPanel/Tree/TreeNode.tsx
@@ -10,7 +10,7 @@ import type { NodeApi } from 'react-arborist';
 import { twMerge } from 'tailwind-merge';
 import NodeIcon from './NodeIcon';
 import { escapeSelector } from '/common/helpers';
-import { MouseAction } from '/common/models';
+import { MouseAction } from '@onlook/types/editor';
 import type { DomElement } from '/common/models/element';
 import type { LayerNode } from '/common/models/element/layers';
